mooved mockAuthentication function outside auth reducer test
