+++ packages/pn-pa-webapp/src/api/notifications/__test__/Notifications.api.test.ts
@@ -9,7 +9,7 @@ import {
   notificationFromBe,
   notificationToFe,
 } from '../../../redux/notification/__test__/test-utils';
-import { mockAuthentication } from '../../../redux/auth/__test__/reducers.test';
+import { mockAuthentication } from '../../../redux/auth/__test__/test-utils';
 import { apiClient, externalClient } from '../../axios';
 import { NotificationsApi } from '../Notifications.api';
 import {
