+++ packages/pn-pa-webapp/src/redux/auth/__test__/test-utils.ts
@@ -1,5 +1,27 @@
+import { AuthApi } from '../../../api/auth/Auth.api';
 import { PartyRole, PNRole } from '../../../models/user';
+import { store } from '../../store';
+import { exchangeToken, logout } from '../actions';
 import { User } from '../types';
+
+export const mockLogin = async (): Promise<any> => {
+  const apiSpy = jest.spyOn(AuthApi, 'exchangeToken');
+  apiSpy.mockResolvedValue(userResponse);
+  return store.dispatch(exchangeToken('mocked-token'));
+};
+
+export const mockLogout = async (): Promise<any> => store.dispatch(logout());
+
+export const mockAuthentication = () => {
+  beforeAll(() => {
+    mockLogin();
+  });
+
+  afterAll(() => {
+    mockLogout();
+    jest.resetAllMocks();
+  });
+};
 
 export const userResponse: User = {
   sessionToken: 'mocked-session-token',
