+++ packages/pn-pa-webapp/src/redux/newNotification/__test__/reducers.test.ts
@@ -3,7 +3,7 @@ import { DigitalDomicileType, PhysicalCommunicationType, RecipientType } from '@
 
 import { NotificationsApi } from '../../../api/notifications/Notifications.api';
 import { PaymentModel } from '../../../models/NewNotification';
-import { mockAuthentication } from '../../auth/__test__/reducers.test';
+import { mockAuthentication } from '../../auth/__test__/test-utils';
 import { store } from '../../store';
 import {
   resetNewNotificationState,
