+++ packages/pn-pa-webapp/src/redux/notification/__test__/reducers.test.ts
@@ -2,7 +2,7 @@ import { resetState } from './../actions';
 import { LegalFactType, NotificationDetail } from '@pagopa-pn/pn-commons';
 
 import { NotificationsApi } from '../../../api/notifications/Notifications.api';
-import { mockAuthentication } from '../../auth/__test__/reducers.test';
+import { mockAuthentication } from '../../auth/__test__/test-utils';
 import { store } from '../../store';
 import {
   getSentNotification,
