+++ packages/pn-personafisica-webapp/src/redux/dashboard/__test__/reducers.test.ts
@@ -8,7 +8,7 @@ import {
 } from '@pagopa-pn/pn-commons';
 
 import { NotificationsApi } from '../../../api/notifications/Notifications.api';
-import { mockAuthentication } from '../../auth/__test__/reducers.test';
+import { mockAuthentication } from '../../auth/__test__/test-utils';
 import { store } from '../../store';
 import {
   getReceivedNotifications,
