+++ packages/pn-personafisica-webapp/src/redux/notification/__test__/reducers.test.ts
@@ -1,6 +1,6 @@
 import { LegalFactType, NotificationDetail, RecipientType } from '@pagopa-pn/pn-commons';
 import { NotificationsApi } from '../../../api/notifications/Notifications.api';
-import { mockAuthentication } from '../../auth/__test__/reducers.test';
+import { mockAuthentication } from '../../auth/__test__/test-utils';
 import { store } from '../../store';
 import {
   getReceivedNotification,
